refactor: extract diagnostic and dataset queries from resumes.ts
Moves listIngestDiagnostics, listArchivedDiagnostics,
listDiagnosticsSourceFacets, countSourceKeyPage,
listWorkflowDatasetPage, and listFieldCoverageDatasetPage
into new resumes_diagnostics.ts module. Resumes.ts reduced
from 2275 to 2036 lines (-10.5%). All extracted functions
re-exported from resumes.ts for backward compatibility.
Web app references updated to new API paths.
